More flexible installer, made codebase worse
Seperated functionality out a bit with methods to allow easier re-use of functions in the future, made the installer not close after any action, and made the installer refresh after a version is wrapper or all Retrowrapper versions are uninstalled. The codebase is now a mess, and uses a ton of bad practices. I'll tidy it up at some point.